지도 넣은 것 불러오는 것 성공

+++ src/main/java/com/example/pamarket00/dto/BoardDto.java
@@ -25,6 +25,6 @@ public class BoardDto {
     int boardLike;
     int boardHitCnt;
     int mapBoardNum;
-    int mapX;
-    int mapY;
+    double mapX;
+    double mapY;
 }
